feat(builder): decouple & modularize src
