+++ src/config.rs
@@ -1,0 +1,3 @@
+pub static FLAKE_TEMPLATE: &'static str = include_str!("assets/flake.nix.template");
+// static SELF_FLAKE_URI: &'static str = "github:tsandrini/flake-parts-builder";
+pub static SELF_FLAKE_URI: &'static str = "."; // TODO only for dev
